Updated Readme
As an absolute newcomer to anything C related I stumbled into some issues due to my own lack of knowledge.
Mainly not being super certain which package to install for the dependency (clang? libc-dev? I know now the difference is irrelevant but would've speed up installation to know so beforehand ^^) and an example on what was meant by `libclang.lib`. I had assumed that to be the file when it was just meant to be the lib directory ith the `libclang.so` file.
In that vein I also added a small example on where you could find that file