Make Opir less strict about unknown struct members
Assimp apparently throws __declspec(dllexport) on structs, what they
hope to achieve with this I'm not entirely sure. But Opir errors out on
this, although maybe rightly so. This is a bit experimental, and it
might cause hard to spot errors in other cases. But Futhark/Opir seems
stable enough to allow a little leeway..