Add warning when overriding type is different size
The type override system exists to allow people to manually wrap parts
of code that Futhark might struggle with. This system also means that if
Nim has an existing type that has the same name as a C defined type then
the nim defined type will take precedence. In #115 it was discovered
that defining `u_int` as `unsigned int` meant that Nims `uint` which is
64 bits long took precedence over the 32 bit long `u_int`. This caused
structures to have incorrect sizes and invalid alignment. To help
against this, and to warn users trying to implement their own types
incorrectly, this warning has been added.